Rename broker core config class
