+++ modules/broker-amqp/src/main/java/io/ballerina/messaging/broker/amqp/Server.java
@@ -32,8 +32,8 @@ import io.ballerina.messaging.broker.amqp.metrics.AmqpMetricManager;
 import io.ballerina.messaging.broker.amqp.metrics.DefaultAmqpMetricManager;
 import io.ballerina.messaging.broker.amqp.metrics.NullAmqpMetricManager;
 import io.ballerina.messaging.broker.auth.AuthManager;
-import io.ballerina.messaging.broker.common.BrokerConfigProvider;
 import io.ballerina.messaging.broker.common.StartupContext;
+import io.ballerina.messaging.broker.common.config.BrokerConfigProvider;
 import io.ballerina.messaging.broker.coordination.BasicHaListener;
 import io.ballerina.messaging.broker.coordination.HaListener;
 import io.ballerina.messaging.broker.coordination.HaStrategy;
