+++ modules/broker-auth/src/main/java/io/ballerina/messaging/broker/auth/authentication/authenticator/impl/JaasAuthenticator.java
@@ -24,8 +24,8 @@ import io.ballerina.messaging.broker.auth.BrokerAuthException;
 import io.ballerina.messaging.broker.auth.authentication.authenticator.Authenticator;
 import io.ballerina.messaging.broker.auth.authentication.sasl.plain.PlainSaslCallbackHandler;
 import io.ballerina.messaging.broker.auth.user.UserStoreManager;
-import io.ballerina.messaging.broker.common.BrokerConfigProvider;
 import io.ballerina.messaging.broker.common.StartupContext;
+import io.ballerina.messaging.broker.common.config.BrokerConfigProvider;
 
 import java.util.Map;
 import javax.security.auth.login.AppConfigurationEntry;
