+++ modules/broker-core/src/test/java/io/ballerina/messaging/broker/core/TopicExchangeTest.java
@@ -21,7 +21,7 @@ package io.ballerina.messaging.broker.core;
 
 import io.ballerina.messaging.broker.common.ValidationException;
 import io.ballerina.messaging.broker.common.data.types.FieldTable;
-import io.ballerina.messaging.broker.core.configuration.BrokerConfiguration;
+import io.ballerina.messaging.broker.core.configuration.BrokerCoreConfiguration;
 import io.ballerina.messaging.broker.core.metrics.NullBrokerMetricManager;
 import io.ballerina.messaging.broker.core.store.dao.impl.NoOpBindingDao;
 import org.testng.Assert;
@@ -61,7 +61,7 @@ public class TopicExchangeTest {
                                                 String publishedTopic) throws BrokerException, ValidationException {
         QueueHandlerFactory factory = new QueueHandlerFactory(null,
                                                               new NullBrokerMetricManager(),
-                                                              new BrokerConfiguration());
+                                                              new BrokerCoreConfiguration());
         QueueHandler handler = factory.createNonDurableQueueHandler(subscribedPattern, false);
         topicExchange.bind(handler, subscribedPattern, FieldTable.EMPTY_TABLE);
 
@@ -79,7 +79,7 @@ public class TopicExchangeTest {
     public void testNegativeSingleTopicMatching(String subscribedPattern,
                                                 String publishedTopic) throws BrokerException, ValidationException {
         QueueHandlerFactory factory = new QueueHandlerFactory(null, new NullBrokerMetricManager(),
-                                                              new BrokerConfiguration());
+                                                              new BrokerCoreConfiguration());
         QueueHandler handler = factory.createNonDurableQueueHandler(subscribedPattern, false);
         topicExchange.bind(handler, subscribedPattern, FieldTable.EMPTY_TABLE);
 
@@ -92,7 +92,7 @@ public class TopicExchangeTest {
     public void testTopicRemoval(String subscribedPattern, String publishedTopic)
             throws BrokerException, ValidationException {
         QueueHandlerFactory factory = new QueueHandlerFactory(null, new NullBrokerMetricManager(),
-                                                              new BrokerConfiguration());
+                                                              new BrokerCoreConfiguration());
         QueueHandler handler = factory.createNonDurableQueueHandler(subscribedPattern, false);
         Queue queue = handler.getQueue();
         topicExchange.bind(handler, subscribedPattern, FieldTable.EMPTY_TABLE);
