+++ modules/broker-metrics/src/main/java/io/ballerina/messaging/broker/metrics/BrokerMetricService.java
@@ -19,8 +19,8 @@
 
 package io.ballerina.messaging.broker.metrics;
 
-import io.ballerina.messaging.broker.common.BrokerConfigProvider;
 import io.ballerina.messaging.broker.common.StartupContext;
+import io.ballerina.messaging.broker.common.config.BrokerConfigProvider;
 import org.wso2.carbon.metrics.core.MetricService;
 import org.wso2.carbon.metrics.core.Metrics;
