+++ modules/broker-metrics/src/test/java/io/ballerina/messaging/broker/metrics/CarbonConfigAdapterTest.java
@@ -19,7 +19,7 @@
 
 package io.ballerina.messaging.broker.metrics;
 
-import io.ballerina.messaging.broker.common.BrokerConfigProvider;
+import io.ballerina.messaging.broker.common.config.BrokerConfigProvider;
 import org.testng.Assert;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
