+++ modules/integration/src/test/java/io/ballerina/messaging/broker/integration/util/TestConfigProvider.java
@@ -19,7 +19,7 @@
 
 package io.ballerina.messaging.broker.integration.util;
 
-import io.ballerina.messaging.broker.common.BrokerConfigProvider;
+import io.ballerina.messaging.broker.common.config.BrokerConfigProvider;
 
 import java.util.HashMap;
 import java.util.Map;
